Client timeout implemented; changed older version of code

+++ utils.h
@@ -21,7 +21,7 @@
 // You may change this if you want to
 struct packet {
     int seqnum;
-    unsigned short acknum;
+    int acknum;
     char ack;
     char last;
     unsigned int length;
@@ -29,7 +29,7 @@ struct packet {
 };
 
 // Utility function to build a packet
-void build_packet(struct packet* pkt, unsigned short seqnum, unsigned short acknum, char last, char ack,unsigned int length, const char* payload) {
+void build_packet(struct packet* pkt, unsigned short seqnum, int acknum, char last, char ack,unsigned int length, const char* payload) {
     pkt->seqnum = seqnum;
     pkt->acknum = acknum;
     pkt->ack = ack;
